backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=4.0.0,<5.0.0",
    "fastapi[standard]>=0.128.7",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
